requirements.txt
build<0.10.0
configparser==6.0.0
GitPython==3.1.40
pytest==7.4.2
pytest-html==4.0.2
python-dateutil==2.8.2
tomli==2.0.1
tox==4.11.3
twine==4.0.2
